управление хорошее, можно переносить в отдельынй класс!

+++ LBEngine/inc/Game.h
@@ -64,7 +64,7 @@ private:
 
     void CreateResources();
 
-    void Update(float deltaTime);
+    void Update(DWORD deltaTime);
     void Render();
     void Clear(const float clearColor[4], float clearDepth, UINT8 clearStencil);
     void Present();
